Docs/Python/requirements.txt
# Manipulació de dades
pandas
numpy

# Connexió a Base de Dades (MongoDB)
pymongo
dnspython

# Machine Learning (K-Means, DBSCAN, Mètriques)
scikit-learn

# Visualització de dades
matplotlib
seaborn

# Variables d'entorn
python-dotenv